+++ Tests/ObserveTests/WhenRunningTestsFromATestObject.swift
@@ -5,6 +5,7 @@
 //  Created by Sam Meech-Ward on 2016-12-09.
 //
 //
+
 @testable import Observe
 import XCTest
 
@@ -25,9 +26,9 @@ class WhenRunningTestsFromATestObject: XCTestCase {
 
     func test_runTest_runsTheTestClosure() {
         var run = false
-        observeTest.closure = {
+        observeTest = ObserveTest(closure: {
             run = true
-        }
+        })
         
         observeTest.runTest()
     
@@ -35,9 +36,9 @@ class WhenRunningTestsFromATestObject: XCTestCase {
     }
     
     func test_runTest_updatesRunningProperties() {
-        observeTest.closure = {
+        observeTest = ObserveTest(closure: {
             XCTAssertTrue(self.observeTest.running)
-        }
+        })
         
         XCTAssertFalse(observeTest.running)
         observeTest.runTest()